Moved PolicyAdapter to types/adapters
