+++ src/services/useVerifyPolicy.ts
@@ -1,7 +1,7 @@
 import { useMutation } from 'react-fetching-library';
 import { DeepPartial } from 'ts-essentials';
 import { Policy } from '../types/Policy';
-import { toServerPolicy } from '../utils/PolicyAdapter';
+import { toServerPolicy } from '../types/adapters/PolicyAdapter';
 import { actionPostPoliciesValidate } from '../generated/ActionCreators';
 
 export const useVerifyPolicyMutation = () => {
