+++ src/types/adapters/__tests__/PolicyAdapter.test.ts
@@ -5,7 +5,7 @@ import {
     toPolicy,
     toServerAction,
     toServerPolicy
-} from '../PolicyAdapter';
+} from '../../types/adapters/PolicyAdapter';
 import {
     NewPolicy,
     PagedServerPolicyResponse,
